requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
gevent==21.1.2
greenlet==1.0.0
gunicorn==20.0.4
idna==2.10
psycopg2==2.8.6
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
zope.event==4.5.0
zope.interface==5.4.0
